requirements.txt
streamlit
duckdb
altair
python-dotenv
scipy
nptdms
notion
notion-client
nptdms==1.7.1
streamlit-authenticator
gdown
PyDrive 
oauth2client
google-auth